src/test/java/org/anne/aoc2023/Day8Test.java
package org.anne.aoc2023;

import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class Day8Test {

    List<String> input = Arrays.asList(
            "RL",
            "",
            "AAA = (BBB, CCC)",
            "BBB = (DDD, EEE)",
            "CCC = (ZZZ, GGG)",
            "DDD = (DDD, DDD)",
            "EEE = (EEE, EEE)",
            "GGG = (GGG, GGG)",
            "ZZZ = (ZZZ, ZZZ)"
    );
    
    List<String> input2 = Arrays.asList(
            "LLR",
            "",
            "AAA = (BBB, BBB)",
            "BBB = (AAA, ZZZ)",
            "ZZZ = (ZZZ, ZZZ)"
    );
    
    List<String> input3 = Arrays.asList(
            "LR",
            "",
            "11A = (11B, XXX)",
            "11B = (XXX, 11Z)",
            "11Z = (11B, XXX)",
            "22A = (22B, XXX)",
            "22B = (22C, 22C)",
            "22C = (22Z, 22Z)",
            "22Z = (22B, 22B)",
            "XXX = (XXX, XXX)"
    );
    
    @Test
    void testPart1() {
        assertEquals(2, Day8.part1(input));
        assertEquals(6, Day8.part1(input2));
    }

    @Test
    void testPart2() {
        assertEquals(6, Day8.part2(input3));
    }
 }
